move sending message logic to Chat.kt
